Update Palindrome.cpp
Removed the cleanInput string in main(). Removed the duplicate call to removePunctuation(convertToLower(inputString)) at line 72; this already happened in the isPalindrome function. Changed the return value at lines 181-185, simply returning the result of a reverseInput == cleanString comparison. Updated some of the names of the variables to account for the names from the assignment instead of my own from scratch!